+++ Patches/SettingsMenuAccessibilityPatch.cs
@@ -365,6 +365,7 @@ internal static class SettingsMenuManager
                 AddOption(AccessibilityOptions.ShortStatusPhrasing, "Short status phrasing");
                 AddOption(AccessibilityOptions.SkipRoutineDecay, "Skip routine status decay");
                 AddOption(AccessibilityOptions.SkipEnemyStatusChanges, "Skip enemy status changes");
+                AddOption(AccessibilityOptions.SpeakChat, "Speak chat messages");
                 AddOption(AccessibilityOptions.DebugMode, "Debug mode");
                 break;
         }